% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{exponential2}
\alias{exponential2}
\title{An exponential relationship to a r.v. x.}
\usage{
exponential2(x, noise, noiseLevel, numNoise, n)
}
\arguments{
\item{x}{A random variable.}

\item{noise}{The total amount of noise to add.}

\item{noiseLevel}{The level of noise.}

\item{numNoise}{The number of levels.}

\item{n}{The length of x.}
}
\value{
The corresponding relationship.
}
\description{
An exponential relationship to a r.v. x.
}

