\name{nlf-package}
\alias{nlf-package}
\alias{nlf}
\docType{package}
\title{
\packageTitle{nlf}
}
\description{
\packageDescription{nlf}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{nlf}
\packageIndices{nlf}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{nlf}

Maintainer: \packageMaintainer{nlf}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
